Move/rename live store configurator
