+++ src/slices/live/live-module-configurator.ts
@@ -1,10 +1,10 @@
 /** @format */
 
 import { PersistConfig, persistReducer, persistStore } from 'redux-persist';
-import { debug } from '../common/debug.js';
-import { IdbPersistStorage } from '../middleware/idb-persist-storage.js';
-import { RootStore, SliceStoreConfigurator, store as globalStore } from '../store.js';
-import { live, LiveState } from './live/live-slice.js';
+import { debug } from '../../common/debug.js';
+import { IdbPersistStorage } from '../../middleware/idb-persist-storage.js';
+import { RootStore, SliceStoreConfigurator, store as globalStore } from '../../store.js';
+import { live, LiveState } from './live-slice.js';
 
 const debugStore = debug('live-store');
