+++ test/unit/components/lineup-game-complete.test.ts
@@ -4,7 +4,7 @@ import '@app/components/lineup-game-complete.js';
 import { LineupGameComplete } from '@app/components/lineup-game-complete.js';
 import { GameDetail, GameStatus } from '@app/models/game.js';
 import { LiveGame } from '@app/models/live.js';
-import { getLiveStoreConfigurator } from '@app/slices/live-store.js';
+import { getLiveStoreConfigurator } from '@app/slices/live/live-module-configurator.js';
 import { RootState, setupStore } from '@app/store.js';
 import { expect, fixture, html } from '@open-wc/testing';
 import { buildGameStateWithCurrentGame } from '../helpers/game-state-setup.js';
