+++ test/unit/components/lineup-game-live.test.ts
@@ -11,8 +11,8 @@ import { FormationType } from '@app/models/formation.js';
 import { GameDetail, GameStatus } from '@app/models/game.js';
 import { LiveGame, LivePlayer, PeriodStatus, getPlayer } from '@app/models/live.js';
 import { PlayerStatus } from '@app/models/player.js';
-import { getLiveStoreConfigurator } from '@app/slices/live-store.js';
 import { endPeriodCreator, selectLiveGameById } from '@app/slices/live/index.js';
+import { getLiveStoreConfigurator } from '@app/slices/live/live-module-configurator.js';
 import { actions as liveActions } from '@app/slices/live/live-slice.js';
 import { RootState, setupStore } from '@app/store.js';
 import { Button } from '@material/mwc-button';
